Configurando os testes.

+++ src/RegiaoIntegrationTest/IntegrationTest.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using RegiaoApi.Context;
@@ -12,35 +13,52 @@ namespace RegiaoIntegrationTest
     {
         protected override IHost CreateHost(IHostBuilder builder)
         {
+            Environment.SetEnvironmentVariable("ASPNETCORE_ENVIRONMENT", "Test");
             var host = base.CreateHost(builder);
+
 
             using (var scope = host.Services.CreateScope())
             {
                 var services = scope.ServiceProvider;
                 var dbContext = services.GetRequiredService<RegiaoDb>();
 
-                dbContext.Database.EnsureDeleted();
-                dbContext.Database.EnsureCreated();
-
-                if (!dbContext.Regioes.Any())
+                if (IsTestEnvironment())
                 {
-                    dbContext.Regioes.AddRange(
-                        new Regiao { Name = "Regiao 1" },
-                        new Regiao { Name = "Regiao 2" },
-                        new Regiao { Name = "Regiao 3" }
-                    );
-                    dbContext.SaveChanges();
+                    dbContext.Database.EnsureDeleted();
+                    dbContext.Database.Migrate();
+
+                    if (!dbContext.Regioes.Any())
+                    {
+                        dbContext.Regioes.AddRange(
+                            new Regiao { Name = "Regiao 1" },
+                            new Regiao { Name = "Regiao 2" },
+                            new Regiao { Name = "Regiao 3" }
+                        );
+                        dbContext.SaveChanges();
+                    }
                 }
             }
-
 
             return host;
         }
 
+        private bool IsTestEnvironment()
+        {
+            return Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") == "Test";
+        }
+
         protected override void ConfigureWebHost(IWebHostBuilder builder)
         {
-            builder.ConfigureServices(services =>
+            builder.ConfigureAppConfiguration((context, config) =>
             {
+                config.AddJsonFile("appsettings.Test.json", optional: false, reloadOnChange: true);
+            });
+
+            builder.ConfigureServices((context, services) =>
+            {
+                var configuration = context.Configuration;
+                var connectionString = configuration.GetConnectionString("DefaultConnection");
+
                 var descriptor = services.SingleOrDefault(d => d.ServiceType == typeof(DbContextOptions<RegiaoDb>));
                 if (descriptor != null)
                 {
@@ -49,10 +67,17 @@ namespace RegiaoIntegrationTest
 
                 services.AddDbContext<RegiaoDb>(options =>
                 {
-                    var connectionString = "Host=localhost;Database=RegiaoTestDb;Username=techuser;Password=techpassword";
                     options.UseNpgsql(connectionString);
                 });
-                //services.AddSingleton<IMessageProducer>(provider => new Producer("test.regiao.updated"));
+
+                var rabbitMqHost = context.Configuration["RabbitMq:Host"];
+                var rabbitMqPort = context.Configuration["RabbitMq:Port"];
+                var rabbitMqUser = context.Configuration["RabbitMq:User"];
+                var rabbitMqPassword = context.Configuration["RabbitMq:Password"];
+                var rabbitMqRegiaoQueue = context.Configuration["RabbitMq:RegiaoQueue"];
+                var rabbitMqDddQueue = context.Configuration["RabbitMq:DddQueue"];
+
+                services.AddSingleton<IMessageProducer>(provider => new Producer(rabbitMqHost, rabbitMqRegiaoQueue, Convert.ToInt32(rabbitMqPort), rabbitMqUser, rabbitMqPassword));
             });
         }
     }
